+++ lib/GenomeFileUtil/GenomeFileUtilImpl.py
@@ -41,7 +41,7 @@ class GenomeFileUtil:
     GenomeFileUtil
 
     Module Description:
-
+    
     '''
 
     ######## WARNING FOR GEVENT USERS ####### noqa
@@ -50,9 +50,9 @@ class GenomeFileUtil:
     # state. A method could easily clobber the state set by another while
     # the latter method is running.
     ######################################### noqa
-    VERSION = "0.8.14"
-    GIT_URL = "https://github.com/kbaseapps/GenomeFileUtil"
-    GIT_COMMIT_HASH = "97a0c33e1e85c5c3351638513e1a5b324d8e5e93"
+    VERSION = "0.9.0"
+    GIT_URL = "https://github.com/kbaseapps/GenomeFileUtil.git"
+    GIT_COMMIT_HASH = "f9a4aaa7ba477336054c4f5b232db75ae52adc84"
 
     #BEGIN_CLASS_HEADER
     #END_CLASS_HEADER
@@ -589,17 +589,18 @@ class GenomeFileUtil:
            scientific_name - will be used to set the scientific name of the
            genome and link to a taxon generate_missing_genes - If the file
            has CDS or mRNA with no corresponding gene, generate a spoofed
-           gene. Off by default) -> structure: parameter "fasta_file" of type
-           "File" -> structure: parameter "path" of String, parameter
-           "shock_id" of String, parameter "ftp_url" of String, parameter
-           "gff_file" of type "File" -> structure: parameter "path" of
+           gene. Off by default existing_assembly_ref - a KBase assembly upa,
+           to associate the metagenome with.) -> structure: parameter
+           "fasta_file" of type "File" -> structure: parameter "path" of
            String, parameter "shock_id" of String, parameter "ftp_url" of
-           String, parameter "genome_name" of String, parameter
-           "workspace_name" of String, parameter "source" of String,
-           parameter "scientific_name" of String, parameter "metadata" of
-           type "usermeta" -> mapping from String to String, parameter
-           "generate_missing_genes" of type "boolean" (A boolean - 0 for
-           false, 1 for true. @range (0, 1))
+           String, parameter "gff_file" of type "File" -> structure:
+           parameter "path" of String, parameter "shock_id" of String,
+           parameter "ftp_url" of String, parameter "genome_name" of String,
+           parameter "workspace_name" of String, parameter "source" of
+           String, parameter "metadata" of type "usermeta" -> mapping from
+           String to String, parameter "generate_missing_genes" of type
+           "boolean" (A boolean - 0 for false, 1 for true. @range (0, 1)),
+           parameter "existing_assembly_ref" of String
         :returns: instance of type "MetagenomeSaveResult" -> structure:
            parameter "metagenome_ref" of String
         """
@@ -632,73 +633,130 @@ class GenomeFileUtil:
         """
         :param params: instance of type "SaveOneGenomeParams" -> structure:
            parameter "workspace" of String, parameter "name" of String,
-           parameter "data" of type "Genome" (Genome object holds much of the
-           data relevant for a genome in KBase Genome publications should be
-           papers about the genome Should the Genome object contain a list of
-           contig_ids too? Source: allowed entries RefSeq, Ensembl,
-           Phytozome, RAST, Prokka, User_upload #allowed entries RefSeq,
-           Ensembl, Phytozome, RAST, Prokka, User_upload controlled
-           vocabulary managed by API Domain is a controlled vocabulary
-           Warnings : mostly controlled vocab but also allow for unstructured
-           Genome_tiers : controlled vocabulary (based on ap input and API
-           checked) Allowed values: #Representative, Reference, ExternalDB,
-           User Examples Tiers: All phytozome - Representative and ExternalDB
-           Phytozome flagship genomes - Reference, Representative and
-           ExternalDB Ensembl - Representative and ExternalDB RefSeq
-           Reference - Reference, Representative and ExternalDB RefSeq
-           Representative - Representative and ExternalDB RefSeq Latest or
-           All Assemblies folder - ExternalDB User Data - User tagged Example
-           Sources: RefSeq, Ensembl, Phytozome, Microcosm, User, RAST,
-           Prokka, (other annotators) @optional warnings contig_lengths
-           contig_ids source_id taxonomy publications @optional
-           ontology_events ontologies_present non_coding_features mrnas
-           genome_type @optional genbank_handle_ref gff_handle_ref
+           parameter "data" of type "Genome" (Genome type -- annotated and
+           assembled genome data. Field descriptions: id - string - KBase
+           legacy data ID scientific_name - string - human readable species
+           name domain - string - human readable phylogenetic domain name
+           (eg. "Bacteria") warnings - list of string - genome-level warnings
+           generated in the annotation process genome_tiers - list of string
+           - controlled vocabulary (based on app input and checked by
+           GenomeFileUtil) A list of labels describing the data source for
+           this genome. Allowed values - Representative, Reference,
+           ExternalDB, User Tier assignments based on genome source: * All
+           phytozome - Representative and ExternalDB * Phytozome flagship
+           genomes - Reference, Representative and ExternalDB * Ensembl -
+           Representative and ExternalDB * RefSeq Reference - Reference,
+           Representative and ExternalDB * RefSeq Representative -
+           Representative and ExternalDB * RefSeq Latest or All Assemblies
+           folder - ExternalDB * User Data - User tagged feature_counts - map
+           of string to integer - total counts of each type of feature keys
+           are a controlled vocabulary of - "CDS", "gene", "misc_feature",
+           "misc_recomb", "mobile_element", "ncRNA" - 72,
+           "non_coding_features", "non_coding_genes",
+           "protein_encoding_gene", "rRNA", "rep_origin", "repeat_region",
+           "tRNA" genetic_code - int - An NCBI-assigned taxonomic category
+           for the organism See here -
+           https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi dna_size
+           - integer - total number of nucleotides num_contigs - integer -
+           total number of contigs in the genome molecule_type - string -
+           controlled vocab - the type of molecule sequenced Possible values
+           are "Unknown", "DNA", "RNA", "genomic DNA", "genomic RNA", "mRNA",
+           "tRNA", "rRNA", "other RNA", "other DNA", "transcribed RNA",
+           "viral cRNA", "unassigned DNA", "unassigned RNA" contig_lengths -
+           list of int - nucleotide length of each contig in the genome
+           Indexes in this list correspond to indexes in the `contig_ids`
+           list. contig_ids - list of str - external database identifiers for
+           each contig (eg. "NC_000913.3") source - str - controlled vocab -
+           descriptor of where this data came from (eg. "RefSeq") Allowed
+           entries RefSeq, Ensembl, Phytozome, RAST, Prokka, User_upload
+           source_id - string - identifier of this genome from the source
+           database (eg. the RefSeq ID such as "NC_000913") md5 - string -
+           checksum of the underlying assembly sequence taxonomy - string -
+           semicolon-delimited taxonomy lineage, in order of parent to child
+           taxon_assignments - mapping of taxonomy namespace to taxon ID.
+           example - {"ncbi": "286", "gtdb": "s__staphylococcus_devriesei"}
+           gc_content - float - ratio of GC count to AT in the genome
+           publications - tuple of (pubmedid, source, title, web_addr, year,
+           authors, journal). See typedef above. ontology_events - A record
+           of the service and method used for a set of ontology assignments
+           on the genome. ontologies_present - a mapping of ontology source
+           id (eg. "GO") to a mapping of term IDs (eg "GO:16209") to term
+           names (eg. "histidine biosynthetic process"). features - array of
+           Feature - protein coding genes (see the separate Feature spec)
+           cdss - array of protein-coding sequences mrnas - array of
+           transcribed messenger RNA sequences (equal to cdss plus 5' and 3'
+           UTRs) non_coding_features - array of features that does not
+           include mRNA, CDS, and protein-encoding genes assembly_ref -
+           workspace reference to an assembly object from which this
+           annotated genome was derived. taxon_ref - workspace reference to a
+           taxon object that classifies the species or strain of this genome.
+           genbank_handle_ref - file server handle reference to the source
+           genbank file for this genome. gff_handle_ref - file server handle
+           reference to the source GFF file for this genome.
+           external_source_origination_date - TODO look at GFU for this
+           release - string - User-supplied release or version of the source
+           data. This most likely will come from an input field in the import
+           app. original_source_file_name - filename from which this genome
+           was derived (eg. genbank or gff filename). notes - TODO
+           quality_scores - TODO suspect - bool - flag of whether this
+           annotation is problematic due to some warning genome_type - string
+           - controlled vocab - One of "draft isolate", "finished isolate",
+           "mag", "sag", "virus", "plasmid", "construct" Features vs. coding
+           sequences: a feature is a sequence in the DNA that codes for a
+           protein, including non-transcribed introns. A coding sequence
+           (stored as `cdss`) includes **only** the sections of the feature
+           that codes for a protein, minus introns and UTRs. @optional
+           warnings contig_lengths contig_ids source_id taxonomy publications
+           @optional ontology_events ontologies_present non_coding_features
+           mrnas genome_type @optional genbank_handle_ref gff_handle_ref
            external_source_origination_date @optional release
            original_source_file_name notes quality_scores suspect
-           assembly_ref @metadata ws gc_content as GC content @metadata ws
-           taxonomy as Taxonomy @metadata ws md5 as MD5 @metadata ws dna_size
-           as Size @metadata ws genetic_code as Genetic code @metadata ws
-           domain as Domain @metadata ws source_id as Source ID @metadata ws
-           source as Source @metadata ws scientific_name as Name @metadata ws
-           genome_type as Genome Type @metadata ws length(features) as Number
-           of Protein Encoding Genes @metadata ws length(cdss) as Number of
-           CDS @metadata ws assembly_ref as Assembly Object @metadata ws
-           num_contigs as Number contigs @metadata ws length(warnings) as
-           Number of Genome Level Warnings @metadata ws suspect as Suspect
-           Genome) -> structure: parameter "id" of type "Genome_id" (KBase
-           genome ID @id kb), parameter "scientific_name" of String,
-           parameter "domain" of String, parameter "warnings" of list of
-           String, parameter "genome_tiers" of list of String, parameter
-           "feature_counts" of mapping from String to Long, parameter
-           "genetic_code" of Long, parameter "dna_size" of Long, parameter
-           "num_contigs" of Long, parameter "molecule_type" of String,
-           parameter "contig_lengths" of list of Long, parameter "contig_ids"
-           of list of String, parameter "source" of String, parameter
-           "source_id" of type "source_id" (Reference to a source_id @id
-           external), parameter "md5" of String, parameter "taxonomy" of
-           String, parameter "gc_content" of Double, parameter "publications"
-           of list of type "publication" (Structure for a publication (float
-           pubmedid string source (ex. Pubmed) string title string web
-           address string  publication year string authors string journal))
-           -> tuple of size 7: parameter "pubmedid" of Double, parameter
-           "source" of String, parameter "title" of String, parameter "url"
-           of String, parameter "year" of String, parameter "authors" of
-           String, parameter "journal" of String, parameter "ontology_events"
-           of list of type "Ontology_event" (@optional ontology_ref
-           method_version eco description) -> structure: parameter "id" of
-           String, parameter "ontology_ref" of type "Ontology_ref" (Reference
-           to a ontology object @id ws KBaseOntology.OntologyDictionary),
-           parameter "method" of String, parameter "method_version" of
-           String, parameter "timestamp" of String, parameter "eco" of
-           String, parameter "description" of String, parameter
-           "ontologies_present" of mapping from String to mapping from String
-           to String, parameter "features" of list of type "Feature"
-           (Structure for a single CDS encoding ???gene??? of a genome ONLY
-           PUT GENES THAT HAVE A CORRESPONDING CDS IN THIS ARRAY NOTE:
-           Sequence is optional. Ideally we can keep it in here, but
-           Recognize due to space constraints another solution may be needed.
-           We may want to add additional fields for other CDM functions
-           (e.g., atomic regulons, coexpressed fids, co_occurring fids,...)
+           assembly_ref @optional taxon_ref taxon_assignments @metadata ws
+           gc_content as GC content @metadata ws taxonomy as Taxonomy
+           @metadata ws md5 as MD5 @metadata ws dna_size as Size @metadata ws
+           genetic_code as Genetic code @metadata ws domain as Domain
+           @metadata ws source_id as Source ID @metadata ws source as Source
+           @metadata ws scientific_name as Name @metadata ws genome_type as
+           Type @metadata ws length(features) as Number of Protein Encoding
+           Genes @metadata ws length(cdss) as Number of CDS @metadata ws
+           assembly_ref as Assembly Object @metadata ws num_contigs as Number
+           contigs @metadata ws length(warnings) as Number of Genome Level
+           Warnings @metadata ws suspect as Suspect Genome) -> structure:
+           parameter "id" of type "Genome_id" (KBase legacy data ID @id kb),
+           parameter "scientific_name" of String, parameter "domain" of
+           String, parameter "warnings" of list of String, parameter
+           "genome_tiers" of list of String, parameter "feature_counts" of
+           mapping from String to Long, parameter "genetic_code" of Long,
+           parameter "dna_size" of Long, parameter "num_contigs" of Long,
+           parameter "molecule_type" of String, parameter "contig_lengths" of
+           list of Long, parameter "contig_ids" of list of String, parameter
+           "source" of String, parameter "source_id" of type "source_id"
+           (Reference to a source_id @id external), parameter "md5" of
+           String, parameter "taxonomy" of String, parameter
+           "taxon_assignments" of mapping from String to String, parameter
+           "gc_content" of Double, parameter "publications" of list of type
+           "publication" (Structure for a publication Elements: (0) pubmedid
+           - float (1) source - string - (ex. Pubmed) (2) title - string (3)
+           string web address - string (4) publication year - string (5)
+           authors - string (6) journal - string) -> tuple of size 7:
+           parameter "pubmedid" of Double, parameter "source" of String,
+           parameter "title" of String, parameter "url" of String, parameter
+           "year" of String, parameter "authors" of String, parameter
+           "journal" of String, parameter "ontology_events" of list of type
+           "Ontology_event" (@optional ontology_ref method_version eco) ->
+           structure: parameter "id" of String, parameter "ontology_ref" of
+           type "Ontology_ref" (Reference to a ontology object @id ws
+           KBaseOntology.OntologyDictionary), parameter "method" of String,
+           parameter "method_version" of String, parameter "timestamp" of
+           String, parameter "eco" of String, parameter "ontologies_present"
+           of mapping from String to mapping from String to String, parameter
+           "features" of list of type "Feature" (Structure for a single CDS
+           encoding ???gene??? of a genome ONLY PUT GENES THAT HAVE A
+           CORRESPONDING CDS IN THIS ARRAY NOTE: Sequence is optional.
+           Ideally we can keep it in here, but Recognize due to space
+           constraints another solution may be needed. We may want to add
+           additional fields for other CDM functions (e.g., atomic regulons,
+           coexpressed fids, co_occurring fids,...)
            protein_translation_length and protein_translation are for longest
            coded protein (representative protein for splice variants) NOTE:
            New Aliases field definitely breaks compatibility. As Does
@@ -720,18 +778,20 @@ class GenomeFileUtil:
            String, parameter "mrnas" of list of String, parameter "children"
            of list of String, parameter "flags" of list of String, parameter
            "warnings" of list of String, parameter "inference_data" of list
-           of type "InferenceInfo" (category;#Maybe a controlled vocabulary
-           type;#Maybe a controlled vocabulary) -> structure: parameter
-           "category" of String, parameter "type" of String, parameter
-           "evidence" of String, parameter "dna_sequence" of String,
-           parameter "dna_sequence_length" of Long, parameter "aliases" of
-           list of tuple of size 2: parameter "fieldname" of String,
-           parameter "alias" of String, parameter "db_xrefs" of list of tuple
-           of size 2: parameter "db_source" of String, parameter
-           "db_identifier" of String, parameter "non_coding_features" of list
-           of type "NonCodingFeature" (Structure for a single feature that is
-           NOT one of the following: Protein encoding gene (gene that has a
-           corresponding CDS) mRNA CDS Note pseudo-genes and Non protein
+           of type "InferenceInfo" (Type spec for the "InferenceInfo" object.
+           TODO docs Found in the `inference_data` fields in mRNAs and CDSs
+           Fields: category - string - TODO type - string - TODO evidence -
+           string - TODO) -> structure: parameter "category" of String,
+           parameter "type" of String, parameter "evidence" of String,
+           parameter "dna_sequence" of String, parameter
+           "dna_sequence_length" of Long, parameter "aliases" of list of
+           tuple of size 2: parameter "fieldname" of String, parameter
+           "alias" of String, parameter "db_xrefs" of list of tuple of size
+           2: parameter "db_source" of String, parameter "db_identifier" of
+           String, parameter "non_coding_features" of list of type
+           "NonCodingFeature" (Structure for a single feature that is NOT one
+           of the following: - Protein encoding gene (gene that has a
+           corresponding CDS) - mRNA - CDS Note pseudo-genes and Non protein
            encoding genes are put into this flags are flag fields in GenBank
            format. This will be a controlled vocabulary. Initially Acceptable
            values are pseudo, ribosomal_slippage, and trans_splicing Md5 is
@@ -748,24 +808,78 @@ class GenomeFileUtil:
            "md5" of String, parameter "parent_gene" of String, parameter
            "children" of list of String, parameter "flags" of list of String,
            parameter "warnings" of list of String, parameter "inference_data"
-           of list of type "InferenceInfo" (category;#Maybe a controlled
-           vocabulary type;#Maybe a controlled vocabulary) -> structure:
-           parameter "category" of String, parameter "type" of String,
-           parameter "evidence" of String, parameter "dna_sequence" of
-           String, parameter "dna_sequence_length" of Long, parameter
-           "aliases" of list of tuple of size 2: parameter "fieldname" of
-           String, parameter "alias" of String, parameter "db_xrefs" of list
-           of tuple of size 2: parameter "db_source" of String, parameter
-           "db_identifier" of String, parameter "cdss" of list of type "CDS"
-           (Structure for a single feature CDS flags are flag fields in
-           GenBank format. This will be a controlled vocabulary. Initially
-           Acceptable values are pseudo, ribosomal_slippage, and
-           trans_splicing Md5 is the md5 of dna_sequence. @optional
-           parent_gene parent_mrna functions ontology_terms note flags
-           warnings @optional inference_data dna_sequence aliases db_xrefs
-           functional_descriptions) -> structure: parameter "id" of type
-           "cds_id" (KBase CDS ID @id external), parameter "location" of list
-           of tuple of size 4: type "Contig_id" (ContigSet contig ID @id
+           of list of type "InferenceInfo" (Type spec for the "InferenceInfo"
+           object. TODO docs Found in the `inference_data` fields in mRNAs
+           and CDSs Fields: category - string - TODO type - string - TODO
+           evidence - string - TODO) -> structure: parameter "category" of
+           String, parameter "type" of String, parameter "evidence" of
+           String, parameter "dna_sequence" of String, parameter
+           "dna_sequence_length" of Long, parameter "aliases" of list of
+           tuple of size 2: parameter "fieldname" of String, parameter
+           "alias" of String, parameter "db_xrefs" of list of tuple of size
+           2: parameter "db_source" of String, parameter "db_identifier" of
+           String, parameter "cdss" of list of type "CDS" (Structure for a
+           single coding sequence. Coding sequences are the sections of a
+           feature's sequence that are translated to a protein (minus introns
+           and UTRs). Fields: id - string - identifier of the coding
+           sequence, such as "b0001_CDS_1" location - list<tuple<string, int,
+           string, int>> - list of locations from where this sequence
+           originates in the original assembly. Each sub-sequence in the list
+           constitutes a section of the resulting CDS. The first element in
+           the tuple corresponds to the "contig_id", such as "NC_000913.3".
+           The second element in the tuple is an index in the contig of where
+           the sequence starts. The third element is either a plus or minus
+           sign indicating whether it is on the 5' to 3' leading strand ("+")
+           or on the 3' to 5' lagging strand ("-"). The last element is the
+           length of the sub-sequence. For a location on the leading strand
+           (denoted by "+"), the index is of the leftmost base, and the
+           sequence extends to the right. For a location on the lagging
+           strand (denoted by "-"), the index is of the rightmost base, and
+           the sequence extends to the left. NOTE: the last element in each
+           tuple is the *length* of each sub-sequence. If you have a location
+           such as ("xyz", 100, "+", 50), then your sequence will go from
+           index 100 to index 149 (this has a length of 50). It *does not* go
+           from index 100 to index 150, as that would have a length of 51.
+           Likewise, if you have the location ("xyz", 100, "-", 50), then the
+           sequence extends from 100 down to 51, which has a length of 50
+           bases. It does not go from index 100 to 50, as that would have a
+           length of 51. md5 - string - md5 of the dna sequence - TODO
+           clarification protein_md5 - string - hash of the protein sequence
+           that this CDS encodes parent_gene - string - gene (feature) from
+           which this CDS comes from, including introns and UTRs that have
+           been removed to create this CDS. parent_mrna - string - mRNA
+           sequence from which this sequence is derived, including UTRs but
+           not introns. note - string - TODO functions - list<string> - list
+           of protein products or chemical processes that this sequence
+           creates, facilitates, or influences. functional_descriptions -
+           list<string> - TODO list of protein products or chemical processes
+           that sequence creates, facilitates, or influences. ontology_terms
+           - mapping<string, mapping<string, list<int>>> - a mapping of
+           ontology source id (eg. "GO") to a mapping of term IDs (eg
+           "GO:16209") to a list of indexes into the ontology_events data
+           (found in the top level of the genome object). The index into an
+           ontology event indicates what service and method created this term
+           assignment. flags - list<string>  - (controlled vocab) fields from
+           the genbank source. A common example is "pseudo" for pseudo-genes
+           that do not encode proteins, which shows up as "/pseudo" in the
+           genbank. Values can be: "pseudo", "ribosomal_slippage",
+           "trans_splicing" warnings - list<string> - TODO inference_data -
+           list<InferenceInfo> - TODO protein_translation - string - amino
+           acid sequence that this CDS gets translated into.
+           protein_translation_length - int - length of the above aliases -
+           list<(string, string)> - alternative list of names or identifiers
+           eg: [["gene", "thrA"], ["locus_tag", "b0002"]] db_xrefs -
+           list<(string, string)> - Identifiers from other databases
+           (database cross-references) The first string is the database name,
+           the second is the database identifier. eg: [["ASAP",
+           "ABE-0000006"], ["EcoGene", "EG11277"]] dna_sequence - string -
+           sequence of exons from the genome that constitute this protein
+           encoding sequence. dna_sequence_length - int - length of the above
+           @optional parent_gene parent_mrna functions ontology_terms note
+           flags warnings @optional inference_data dna_sequence aliases
+           db_xrefs functional_descriptions) -> structure: parameter "id" of
+           type "cds_id" (KBase CDS ID @id external), parameter "location" of
+           list of tuple of size 4: type "Contig_id" (ContigSet contig ID @id
            external), Long, String, Long, parameter "md5" of String,
            parameter "protein_md5" of String, parameter "parent_gene" of type
            "Feature_id" (KBase Feature ID @id external), parameter
@@ -775,20 +889,69 @@ class GenomeFileUtil:
            parameter "ontology_terms" of mapping from String to mapping from
            String to list of Long, parameter "flags" of list of String,
            parameter "warnings" of list of String, parameter "inference_data"
-           of list of type "InferenceInfo" (category;#Maybe a controlled
-           vocabulary type;#Maybe a controlled vocabulary) -> structure:
-           parameter "category" of String, parameter "type" of String,
-           parameter "evidence" of String, parameter "protein_translation" of
-           String, parameter "protein_translation_length" of Long, parameter
-           "aliases" of list of tuple of size 2: parameter "fieldname" of
-           String, parameter "alias" of String, parameter "db_xrefs" of list
-           of tuple of size 2: parameter "db_source" of String, parameter
-           "db_identifier" of String, parameter "dna_sequence" of String,
-           parameter "dna_sequence_length" of Long, parameter "mrnas" of list
-           of type "mRNA" (Structure for a single feature mRNA flags are flag
-           fields in GenBank format. This will be a controlled vocabulary.
-           Initially Acceptable values are pseudo, ribosomal_slippage, and
-           trans_splicing Md5 is the md5 of dna_sequence. @optional
+           of list of type "InferenceInfo" (Type spec for the "InferenceInfo"
+           object. TODO docs Found in the `inference_data` fields in mRNAs
+           and CDSs Fields: category - string - TODO type - string - TODO
+           evidence - string - TODO) -> structure: parameter "category" of
+           String, parameter "type" of String, parameter "evidence" of
+           String, parameter "protein_translation" of String, parameter
+           "protein_translation_length" of Long, parameter "aliases" of list
+           of tuple of size 2: parameter "fieldname" of String, parameter
+           "alias" of String, parameter "db_xrefs" of list of tuple of size
+           2: parameter "db_source" of String, parameter "db_identifier" of
+           String, parameter "dna_sequence" of String, parameter
+           "dna_sequence_length" of Long, parameter "mrnas" of list of type
+           "mRNA" (The mRNA is the transcribed sequence from the original
+           feature, minus the introns, but including the UTRs. Fields: id -
+           string - identifying string for the mRNA location -
+           list<tuple<string, int, string, int>> - list of locations from
+           where this sequence originates in the original assembly. Each
+           sub-sequence in the list constitutes a section of the resulting
+           CDS. The first element in the tuple corresponds to the
+           "contig_id", such as "NC_000913.3". The second element in the
+           tuple is an index in the contig of where the sequence starts. The
+           third element is either a plus or minus sign indicating whether it
+           is on the 5' to 3' leading strand ("+") or on the 3' to 5' lagging
+           strand ("-"). The last element is the length of the sub-sequence.
+           For a location on the leading strand (denoted by "+"), the index
+           is of the leftmost base, and the sequence extends to the right.
+           For a location on the lagging strand (denoted by "-"), the index
+           is of the rightmost base, and the sequence extends to the left.
+           NOTE: the last element in each tuple is the *length* of each
+           sub-sequence. If you have a location such as ("xyz", 100, "+",
+           50), then your sequence will go from index 100 to index 149 (this
+           has a length of 50). It *does not* go from index 100 to index 150,
+           as that would have a length of 51. Likewise, if you have the
+           location ("xyz", 100, "-", 50), then the sequence extends from 100
+           down to 51, which has a length of 50 bases. It does not go from
+           index 100 to 50, as that would have a length of 51. md5 - string -
+           md5 of the dna sequence - TODO clarification parent_gene -
+           Feature_id - corresponding feature for this sequence, including
+           introns and UTRs cds - string - corresponding coding sequence for
+           this mRNA (the sequence minus UTRs) dna_sequence - string -
+           sequence of UTRs and exons from the genome that constitute this
+           mRNA dna_sequence_length - int - length of the above note - string
+           - TODO functions - list<string> - TODO list of protein products or
+           chemical processes that sequence creates, facilitates, or
+           influences. functional_descriptions - list<string> - TODO list of
+           protein products or chemical processes that sequence creates,
+           facilitates, or influences. ontology_terms - mapping<string,
+           mapping<string, list<int>>> - a mapping of ontology source id (eg.
+           "GO") to a mapping of term IDs (eg "GO:16209") to a list of
+           indexes into the ontology_events data (found in the top level of
+           the genome object). The index into an ontology event indicates
+           what service and method created this term assignment. flags -
+           list<string> - controlled vocab - fields from the genbank source.
+           A common example is "pseudo" for pseudo-genes that do not encode
+           proteins, which shows up as "/pseudo" in the genbank. Values can
+           be: "pseudo", "ribosomal_slippage", "trans_splicing" warnings -
+           list<string> - TODO inference_data - list<InferenceInfo> - TODO
+           aliases - list<(string, string)> - alternative list of names or
+           identifiers eg: [["gene", "thrA"], ["locus_tag", "b0002"]]
+           db_xrefs - list<(string, string)> - Identifiers from other
+           databases (database cross-references). The first string is the
+           database name, the second is the database identifier. eg:
+           [["ASAP", "ABE-0000006"], ["EcoGene", "EG11277"]] @optional
            parent_gene cds functions ontology_terms note flags warnings
            @optional inference_data dna_sequence aliases db_xrefs
            functional_descriptions) -> structure: parameter "id" of type
@@ -804,37 +967,41 @@ class GenomeFileUtil:
            "ontology_terms" of mapping from String to mapping from String to
            list of Long, parameter "flags" of list of String, parameter
            "warnings" of list of String, parameter "inference_data" of list
-           of type "InferenceInfo" (category;#Maybe a controlled vocabulary
-           type;#Maybe a controlled vocabulary) -> structure: parameter
-           "category" of String, parameter "type" of String, parameter
-           "evidence" of String, parameter "aliases" of list of tuple of size
-           2: parameter "fieldname" of String, parameter "alias" of String,
-           parameter "db_xrefs" of list of tuple of size 2: parameter
-           "db_source" of String, parameter "db_identifier" of String,
-           parameter "assembly_ref" of type "Assembly_ref" (Reference to an
-           Assembly object in the workspace @id ws
-           KBaseGenomeAnnotations.Assembly), parameter "taxon_ref" of type
-           "Taxon_ref" (Reference to a taxon object @id ws
-           KBaseGenomeAnnotations.Taxon), parameter "genbank_handle_ref" of
-           type "genbank_handle_ref" (Reference to a handle to the Genbank
-           file on shock @id handle), parameter "gff_handle_ref" of type
-           "gff_handle_ref" (Reference to a handle to the GFF file on shock
-           @id handle), parameter "external_source_origination_date" of
-           String, parameter "release" of String, parameter
-           "original_source_file_name" of String, parameter "notes" of
-           String, parameter "quality_scores" of list of type
-           "GenomeQualityScore" (Score_interpretation : fraction_complete -
-           controlled vocabulary managed by API @optional method_report_ref
-           method_version) -> structure: parameter "method" of String,
-           parameter "method_report_ref" of type "Method_report_ref"
-           (Reference to a report object @id ws KBaseReport.Report),
-           parameter "method_version" of String, parameter "score" of String,
-           parameter "score_interpretation" of String, parameter "timestamp"
-           of String, parameter "suspect" of type "Bool", parameter
-           "genome_type" of String, parameter "hidden" of type "boolean" (A
-           boolean - 0 for false, 1 for true. @range (0, 1)), parameter
-           "upgrade" of type "boolean" (A boolean - 0 for false, 1 for true.
-           @range (0, 1))
+           of type "InferenceInfo" (Type spec for the "InferenceInfo" object.
+           TODO docs Found in the `inference_data` fields in mRNAs and CDSs
+           Fields: category - string - TODO type - string - TODO evidence -
+           string - TODO) -> structure: parameter "category" of String,
+           parameter "type" of String, parameter "evidence" of String,
+           parameter "aliases" of list of tuple of size 2: parameter
+           "fieldname" of String, parameter "alias" of String, parameter
+           "db_xrefs" of list of tuple of size 2: parameter "db_source" of
+           String, parameter "db_identifier" of String, parameter
+           "assembly_ref" of type "Assembly_ref" (Reference to an Assembly
+           object in the workspace @id ws KBaseGenomeAnnotations.Assembly),
+           parameter "taxon_ref" of type "Taxon_ref" (Reference to a taxon
+           object @id ws KBaseGenomeAnnotations.Taxon), parameter
+           "genbank_handle_ref" of type "genbank_handle_ref" (Reference to a
+           handle to the Genbank file on shock @id handle), parameter
+           "gff_handle_ref" of type "gff_handle_ref" (Reference to a handle
+           to the GFF file on shock @id handle), parameter
+           "external_source_origination_date" of String, parameter "release"
+           of String, parameter "original_source_file_name" of String,
+           parameter "notes" of String, parameter "quality_scores" of list of
+           type "GenomeQualityScore" (Genome quality score Fields: method -
+           string - TODO method_report_ref - string - TODO method_version -
+           string - TODO score: string - TODO score_interpretation - string -
+           TODO timestamp - string - TODO Score_interpretation -
+           fraction_complete - controlled vocabulary managed by API @optional
+           method_report_ref method_version) -> structure: parameter "method"
+           of String, parameter "method_report_ref" of type
+           "Method_report_ref" (Reference to a report object @id ws
+           KBaseReport.Report), parameter "method_version" of String,
+           parameter "score" of String, parameter "score_interpretation" of
+           String, parameter "timestamp" of String, parameter "suspect" of
+           type "Bool", parameter "genome_type" of String, parameter "hidden"
+           of type "boolean" (A boolean - 0 for false, 1 for true. @range (0,
+           1)), parameter "upgrade" of type "boolean" (A boolean - 0 for
+           false, 1 for true. @range (0, 1))
         :returns: instance of type "SaveGenomeResult" -> structure: parameter
            "info" of type "object_info" (Information about an object,
            including user provided metadata. obj_id objid - the numerical id
@@ -892,6 +1059,116 @@ class GenomeFileUtil:
         # At some point might do deeper type checking...
         if not isinstance(returnVal, dict):
             raise ValueError('Method save_one_genome return value ' +
+                             'returnVal is not type dict as required.')
+        # return the results
+        return [returnVal]
+
+    def ws_obj_gff_to_genome(self, ctx, params):
+        """
+        :param params: instance of type "WsObjGFFToGenomeParams" ->
+           structure: parameter "ws_ref" of String, parameter "gff_file" of
+           type "File" -> structure: parameter "path" of String, parameter
+           "shock_id" of String, parameter "ftp_url" of String, parameter
+           "genome_name" of String, parameter "workspace_name" of String,
+           parameter "source" of String, parameter "taxon_wsname" of String,
+           parameter "taxon_id" of String, parameter "release" of String,
+           parameter "genetic_code" of Long, parameter "scientific_name" of
+           String, parameter "metadata" of type "usermeta" -> mapping from
+           String to String, parameter "generate_missing_genes" of type
+           "boolean" (A boolean - 0 for false, 1 for true. @range (0, 1))
+        :returns: instance of type "GenomeSaveResult" -> structure: parameter
+           "genome_ref" of String
+        """
+        # ctx is the context object
+        # return variables are: returnVal
+        #BEGIN ws_obj_gff_to_genome
+        au = AssemblyUtil(self.cfg.callbackURL)
+        ws = Workspace(url=self.cfg.workspaceURL)
+
+        if params.get('ws_ref'):
+            input_ref = params['ws_ref']
+        else:
+            raise ValueError(f"ws_ref argument required for ws_obj_gff_to_metagenome method") 
+        # input should only be of workspace object types:
+        #   - 'KBaseGenomes.Genome'
+        #   - 'KBaseGenomeAnnotations.Assembly'
+        #   - 'KBaseGenomes.ContigSet'
+        obj_type = ws.get_object_info3({'objects': [{'ref': input_ref}]})['infos'][0][2]
+        obj_type = obj_type.split('-')[0]
+        valid_types = ["KBaseGenomes.Genome", "KBaseGenomes.ContigSet", "KBaseGenomeAnnotations.Assembly"]
+        if obj_type not in valid_types:
+            raise ValueError(f"input reference must be of type KBaseGenomes.Genome or "
+                             f"KBaseGenomes.ContigSet or KBaseGenomeAnnotations.Assembly "
+                             f"not {obj_type}")
+
+        fasta_file = [val['paths'][0] for key, val in au.get_fastas({'ref_lst': [input_ref]}).items()][0]
+
+        params['fasta_file'] = {
+            'path': fasta_file
+        }
+        params['use_existing_assembly'] = input_ref
+        importer = FastaGFFToGenome(self.cfg)
+        returnVal = importer.import_file(params)
+
+        #END ws_obj_gff_to_genome
+
+        # At some point might do deeper type checking...
+        if not isinstance(returnVal, dict):
+            raise ValueError('Method ws_obj_gff_to_genome return value ' +
+                             'returnVal is not type dict as required.')
+        # return the results
+        return [returnVal]
+
+    def ws_obj_gff_to_metagenome(self, ctx, params):
+        """
+        :param params: instance of type "WsObjGFFToMetagenomeParams" ->
+           structure: parameter "ws_ref" of String, parameter "gff_file" of
+           type "File" -> structure: parameter "path" of String, parameter
+           "shock_id" of String, parameter "ftp_url" of String, parameter
+           "genome_name" of String, parameter "workspace_name" of String,
+           parameter "source" of String, parameter "metadata" of type
+           "usermeta" -> mapping from String to String, parameter
+           "generate_missing_genes" of type "boolean" (A boolean - 0 for
+           false, 1 for true. @range (0, 1))
+        :returns: instance of type "MetagenomeSaveResult" -> structure:
+           parameter "metagenome_ref" of String
+        """
+        # ctx is the context object
+        # return variables are: returnVal
+        #BEGIN ws_obj_gff_to_metagenome
+        au = AssemblyUtil(self.cfg.callbackURL)
+        ws = Workspace(url=self.cfg.workspaceURL)
+
+        if params.get('ws_ref'):
+            input_ref = params['ws_ref']
+        else:
+            raise ValueError(f"ws_ref argument required for ws_obj_gff_to_metagenome method")
+        # input should only be of workspace object types:
+        #   - 'KBaseMetagenomes.AnnotatedMetagenomeAssembly'
+        #   - 'KBaseGenomeAnnotations.Assembly'
+        #   - 'KBaseGenomes.ContigSet'
+        obj_type = ws.get_object_info3({'objects': [{'ref': input_ref}]})['infos'][0][2]
+        obj_type = obj_type.split('-')[0]
+        valid_types = ["KBaseMetagenomes.AnnotatedMetagenomeAssembly", "KBaseGenomes.ContigSet", "KBaseGenomeAnnotations.Assembly"]
+        if obj_type not in valid_types:
+            raise ValueError(f"input reference must be of type KBaseMetagenomes.AnnotatedMetagenomeAssembly or "
+                             f"KBaseGenomes.ContigSet or KBaseGenomeAnnotations.Assembly "
+                             f"not {obj_type}")
+
+        fasta_file = [val['paths'][0] for key, val in au.get_fastas({'ref_lst': [input_ref]}).items()][0]
+        params['fasta_file'] = {
+          'path': fasta_file
+        }
+        params['use_existing_assembly'] = input_ref
+        params['is_metagenome'] = True
+
+        importer = FastaGFFToGenome(self.cfg)
+        returnVal = importer.import_file(params)
+        #END ws_obj_gff_to_metagenome
+
+        # At some point might do deeper type checking...
+        if not isinstance(returnVal, dict):
+            raise ValueError('Method ws_obj_gff_to_metagenome return value ' +
                              'returnVal is not type dict as required.')
         # return the results
         return [returnVal]