+++ test/core/gff_up_down_up_test.py
@@ -32,8 +32,8 @@ class GenomeFileUtilTest(unittest.TestCase):
         cls.wsURL = cls.cfg['workspace-url']
         cls.wsClient = workspaceService(cls.wsURL, token=token)
         cls.serviceImpl = GenomeFileUtil(cls.cfg)
-        gff_path = "data/fasta_gff/RefSeq/Bacterial_Data/NC_021490.gff.gz"
-        fasta_path = "data/fasta_gff/RefSeq/Bacterial_Data/NC_021490.fasta.gz"
+        cls.gff_path = "data/fasta_gff/RefSeq/Bacterial_Data/NC_021490.gff.gz"
+        cls.fasta_path = "data/fasta_gff/RefSeq/Bacterial_Data/NC_021490.fasta.gz"
         suffix = int(time.time() * 1000)
         cls.wsName = "test_GenomeFileUtil_" + str(suffix)
         cls.wsClient.create_workspace({'workspace': cls.wsName})
@@ -43,12 +43,13 @@ class GenomeFileUtilTest(unittest.TestCase):
             'workspace_name': cls.wsName,
             'genome_name': 'MyGenome',
             'taxon_id': '243276',
-            'fasta_file': {'path': fasta_path},
-            'gff_file': {'path': gff_path},
+            'fasta_file': {'path': cls.fasta_path},
+            'gff_file': {'path': cls.gff_path},
             'source': 'GFF',
             'type': 'Reference'
         })[0]
         data_file_cli = DataFileUtil(os.environ['SDK_CALLBACK_URL'])
+        cls.genome_ref = result['genome_ref']
         cls.genome_orig = data_file_cli.get_objects(
             {'object_refs': [result['genome_ref']]})['data'][0]['data']
 
@@ -60,7 +61,7 @@ class GenomeFileUtilTest(unittest.TestCase):
         new_result = cls.serviceImpl.fasta_gff_to_genome(cls.ctx, {
             'workspace_name': cls.wsName,
             'genome_name': 'MyGenome',
-            'fasta_file': {'path': fasta_path},
+            'fasta_file': {'path': cls.fasta_path},
             'gff_file': {'path': down_result['file_path']},
             'source': 'GFF',
             'type': 'Reference',
@@ -74,9 +75,7 @@ class GenomeFileUtilTest(unittest.TestCase):
             cls.wsClient.delete_workspace({'workspace': cls.wsName})
             print('Test workspace was deleted')
 
-    def feature_list_comparison(self, feature_list_name):
-        genome_orig = self.genome_orig
-        genome_new = self.genome_new
+    def feature_list_comparison(self, genome_orig, genome_new, feature_list_name):
         self.assertTrue(len(genome_orig[feature_list_name]) == len(genome_new[feature_list_name]),
                         feature_list_name + " list is not of equal length in Original and New Genomes.")
         print("\n\n" + feature_list_name + " TOTAL NUMBER:" + str(len(genome_orig[feature_list_name])))
@@ -132,14 +131,31 @@ class GenomeFileUtilTest(unittest.TestCase):
                          "There were %d first pass matches and %d second pass matches out of %d items in %s" %
                          (first_pass_matches, second_pass_matches, len(orig_dict), feature_list_name))
 
+    def test_genome_gff_to_genome(self):
+        result = self.serviceImpl.ws_obj_gff_to_genome(self.ctx, {
+            'workspace_name': self.wsName,
+            'genome_name': 'MyGenome',
+            'gff_file': {'path': self.gff_path},
+            'ws_ref': self.genome_ref,
+            'source': 'GFF',
+            'type': 'Reference',
+            'taxon_id': '243276'
+        })[0]
+        dfu = DataFileUtil(os.environ['SDK_CALLBACK_URL'])
+        genome = dfu.get_objects({"object_refs": [result['genome_ref']]})['data'][0]['data']
+        self.feature_list_comparison(self.genome_orig, genome, 'features')
+        self.feature_list_comparison(self.genome_orig, genome, 'cdss')
+        self.feature_list_comparison(self.genome_orig, genome, 'mrnas')
+        self.feature_list_comparison(self.genome_orig, genome, 'non_coding_features')
+
     def test_gene_features(self):
-        self.feature_list_comparison("features")
+        self.feature_list_comparison(self.genome_orig, self.genome_new, "features")
 
     def test_cds_features(self):
-        self.feature_list_comparison("cdss")
+        self.feature_list_comparison(self.genome_orig, self.genome_new, "cdss")
 
     def test_mrna_features(self):
-        self.feature_list_comparison("mrnas")
+        self.feature_list_comparison(self.genome_orig, self.genome_new, "mrnas")
 
     def test_ncf_features(self):
-        self.feature_list_comparison("non_coding_features")
+        self.feature_list_comparison(self.genome_orig, self.genome_new, "non_coding_features")