+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.2] - 2020-02-11
+
+### Fixed
+
+- keyError: "warnings" fix for Metagenome upload
+
+## [0.10.1] - 2020-02-10
+
+### Fixed
+
+- ws_obj_gff_to_metagenome was erroneously outputting non-metagenome filestypes when the "is_metagenome" parameter was not specified.
+- protein sequences were still sometimes saved for some metagenome objects within the parent gene of a cds. protein_translation no longer saved in metagenome.
+
+## [0.10.0] - 2020-02-05
+
+### Changed
+
+- Adding two new functions: ws_obj_gff_to_metagenome, ws_obj_gff_to_genome
+
 ## [0.9.0] - 2019-10-24
 
 ### Changed